rename : Recommendation 디렉토리 밑 entity -> domain 으로 변경
