+++ src/main/java/com/smartcar/dororok/recommendation/service/RecommendationService.java
@@ -3,14 +3,12 @@ package com.smartcar.dororok.recommendation.service;
 import com.smartcar.dororok.global.auth.utils.SecurityUtils;
 import com.smartcar.dororok.location.dto.LocationDto;
 import com.smartcar.dororok.location.service.LocationService;
-import com.smartcar.dororok.member.domain.entitiy.FavoriteGenres;
-import com.smartcar.dororok.member.domain.entitiy.Genre;
 import com.smartcar.dororok.member.domain.entitiy.Member;
 import com.smartcar.dororok.member.repository.FavoriteGenresRepository;
 import com.smartcar.dororok.member.repository.MemberRepository;
-import com.smartcar.dororok.recommendation.entity.MusicMode;
-import com.smartcar.dororok.recommendation.entity.dto.MusicRecommendationDto;
-import com.smartcar.dororok.recommendation.entity.dto.MusicRecommendationToDjangoDto;
+import com.smartcar.dororok.recommendation.domain.MusicMode;
+import com.smartcar.dororok.recommendation.domain.dto.MusicRecommendationDto;
+import com.smartcar.dororok.recommendation.domain.dto.MusicRecommendationToDjangoDto;
 import com.smartcar.dororok.weather.domain.GetWeatherDto;
 import com.smartcar.dororok.weather.service.WeatherService;
 import lombok.RequiredArgsConstructor;
